requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.0.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
auto-gptq==0.5.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bangla==0.0.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitarray==2.8.2
bleach==6.0.0
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
boltons==23.0.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
cmake==3.27.5
coloredlogs==15.0.1
comm==0.1.4
constants==0.6.0
constraints==0.9.5
contourpy==1.1.1
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.30
dataclasses-json==0.5.14
datasets==2.14.6
dateparser==1.1.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
docarray==0.39.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
emoji==2.8.0
encodec==0.1.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.2
fastjsonschema==2.18.0
filelock==3.12.4
filetype==1.2.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
g2pkk==0.1.2
gekko==1.0.6
gitdb==4.0.10
GitPython==3.1.37
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-generativeai==0.2.2
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
gpt4all==2.0.1
greenlet==3.0.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.0
httpx==0.25.2
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflect==5.6.0
InstructorEmbedding==1.0.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==1.2.0
jsonmerge==1.9.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
k-diffusion==0.0.16
kiwisolver==1.4.5
kornia==0.7.0
langchain==0.0.314
langdetect==1.0.9
langsmith==0.0.43
lazy_loader==0.3
librosa==0.10.0
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
num2words==0.5.12
numba==0.57.0
numexpr==2.8.7
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
orjson==3.9.9
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pdfreader==0.1.12
peft==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
posthog==3.0.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.11
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.16.1
Pympler==1.0.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pynndescent==0.5.10
pyparsing==3.1.1
pypdf==3.16.4
PyPDF2==3.0.1
PyPika==0.48.9
pypinyin==0.49.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-json-logger==2.0.7
python-magic==0.4.27
python-telegram-bot==20.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rapidfuzz==3.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rouge==1.0.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
semver==3.0.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
SpeechRecognition==3.10.0
SQLAlchemy==2.0.22
stack-data==0.6.2
starlette==0.27.0
streamlit==1.18.1
streamlit-autorefresh==1.0.1
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.18
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
trainer==0.0.31
traitlets==5.10.0
trampoline==0.1.2
transformers==4.31.0
triton==2.0.0
TTS==0.17.4
typer==0.9.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
umap-learn==0.5.1
Unidecode==1.3.6
unstructured==0.10.22
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
vertexai==0.0.1
wandb==0.15.11
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
